+++ core/src/main/java/com/afs/dragbits/Autos/AutoJugador.java
@@ -1,14 +1,10 @@
 package com.afs.dragbits.autos;
 
-import com.badlogic.gdx.graphics.Color;
-
-/**
- * Representa el auto controlado por el jugador.
- */
+/**auto controlado por el jugador, carga el sprite sheet auto 1-sheet.png */
 public class AutoJugador extends Auto {
 
     public AutoJugador(float posX, float posY) {
-        // (posX, posY, velocidadMaxima, aceleracion, Color)
-        super(posX, posY, 400f, 150f, Color.RED); // Auto del jugador de color Rojo
+        // (posX, posY, velocidadMaxima, aceleracion, rutaSpriteSheet)
+        super(posX, posY, 400f, 150f, "sprites/auto 1-sheet.png");
     }
 }