+++ CHANGELOG.md
@@ -8,3 +8,15 @@
 - Creación del archivo README.md con la información del proyecto.
 - Creación del archivo CHANGELOG.md para el registro de modificaciones.
 - Creación y confifuración de la wiki de GitHub
+
+### Added 5/8/2026
+
+- Version Inicial del control de aceleracion y cambio
+- HUD simplificado por texto para mostrar informacion basica 
+- Version preliminar de la clase Auto y AutoJugador con velocidad maxima, aceleracion y rpm maximas
+- Diseño simplificado de la pista y del auto
+
+
+### Updated 10/8/2026
+
+- Diseño basico de un primer auto con spritesheet, y reaciona segun la situacion (cambio, aceleracion, estatico).